% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickset-package.R
\docType{package}
\name{brickset-package}
\alias{brickset-package}
\title{R package to interface with the Brickset API for getting data about LEGO.}
\description{
R package to interface with the Brickset API for getting data about LEGO.
}
\author{
\email{jason@bryer.org}
}
\keyword{data}
\keyword{lego}
