% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_reviews.R
\name{download_reviews}
\alias{download_reviews}
\title{Downloads a pre-built reviews data frame.}
\usage{
download_reviews(dest = tempdir(), ...)
}
\arguments{
\item{dest}{directory to download the \code{reviews.rda} file to.}

\item{...}{other parameters passed to \code{\link[piggyback:pb_download]{piggyback::pb_download()}}.}
}
\value{
a data frame with all the reviews as of the tag date.
}
\description{
This function will return a data frame with all the reviews as of last package
build. Since this data frame is larger than what is allowed in CRAN packages,
the data files are saved as releases on Github. The \code{\link[piggyback:pb_upload]{piggyback::pb_upload()}}
function is used to upload the data frame.
}
\details{
To see what versions are available use the \code{piggyback::pb_list(repo = 'jbryer/brickset')}
function call. By default the latest version will be returned. For reproducibility
you can use the \code{tag} parameter to return a specific version of the data frame.
}
