% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInstructions.R
\name{getInstructions}
\alias{getInstructions}
\title{Download list of instructions for given set.}
\usage{
getInstructions(
  setID,
  setNumber,
  key,
  username,
  password,
  userHash = login(username, password, key)
)
}
\arguments{
\item{setNumber}{the set number from on the LEGO box}

\item{key}{the Brickset key.}

\item{username}{the Brickset username.}

\item{password}{the Brickset password.}

\item{userHash}{the hash used for a logged in user. See \link{login}.}

\item{setId}{the ID of the set (see \code{data(legosets)})}
}
\value{
a data.frame with the reviews.
}
\description{
Download list of instructions for given set.
}
