% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeyUsageStats.R
\name{getKeyUsageStats}
\alias{getKeyUsageStats}
\title{Check the Brickset API key.}
\usage{
getKeyUsageStats(key, error = TRUE)
}
\arguments{
\item{key}{the API key}

\item{error}{if TRUE, the function will throw an error. If FALSE, then the
function will return TRUE if the key check failed.}
}
\value{
TRUE if the API key is fine.
}
\description{
You can request an API key on the Brickset website here:
http://brickset.com/tools/webservices/requestkey
}
