% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSets.R
\name{getSets}
\alias{getSets}
\title{Downloads LEGO data from Brickset.}
\usage{
getSets(year, key = getOption("brickset_key"), ...)
}
\arguments{
\item{year}{the year of data to download.}

\item{key}{the Brickset key.}

\item{...}{other parameters passed to \code{\link{getUserHash}} including
the Brickset username and password if they are not available from
\code{getOption('brickset_username')} and \code{getOption('brickset_password')}.}
}
\value{
a data.frame with all sets from the given year.
\describe{
\item{setID}{integer; 21546 unique values}
\item{number}{character; 20037 unique values}
\item{numberVariant}{integer; 25 unique values}
\item{name}{character; 18081 unique values}
\item{year}{integer; 56 unique values}
\item{theme}{character; 171 unique values}
\item{themeGroup}{character; 17 unique values}
\item{subtheme}{character; 1055 unique values}
\item{category}{character; 7 unique values}
\item{released}{logical; 2 unique values}
\item{pieces}{integer; 1621 unique values}
\item{minifigs}{integer; 35 unique values}
\item{bricksetURL}{character; 21546 unique values}
\item{rating}{numeric; 29 unique values}
\item{reviewCount}{integer; 64 unique values}
\item{packagingType}{character; 19 unique values}
\item{availability}{character; 12 unique values}
\item{agerange_min}{integer; 19 unique values}
\item{thumbnailURL}{character; 20514 unique values}
\item{imageURL}{character; 20514 unique values}
\item{US_retailPrice}{numeric; 194 unique values}
\item{US_dateFirstAvailable}{Date; 1213 unique values}
\item{US_dateLastAvailable}{Date; 2487 unique values}
\item{UK_retailPrice}{numeric; 228 unique values}
\item{UK_dateFirstAvailable}{Date; 1132 unique values}
\item{UK_dateLastAvailable}{Date; 2393 unique values}
\item{CA_retailPrice}{numeric; 204 unique values}
\item{CA_dateFirstAvailable}{Date; 962 unique values}
\item{CA_dateLastAvailable}{Date; 2165 unique values}
\item{DE_retailPrice}{numeric; 183 unique values}
\item{DE_dateFirstAvailable}{Date; 724 unique values}
\item{DE_dateLastAvailable}{Date; 1590 unique values}
\item{height}{numeric; 292 unique values}
\item{width}{numeric; 349 unique values}
\item{depth}{numeric; 318 unique values}
\item{weight}{numeric; 1296 unique values}
}
}
\description{
Brickset API documentation is available here:
https://brickset.com/article/52664/api-version-3-documentation/
}
