% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubthemes.R
\name{getSubthemes}
\alias{getSubthemes}
\title{Returns a table of sub-themes for a given theme with number of sets and years active.}
\usage{
getSubthemes(theme, key = getOption("brickset_key"), ...)
}
\arguments{
\item{theme}{the theme}

\item{...}{other parameters passed to \code{\link{getUserHash}} including
the Brickset username and password if they are not available from
\code{getOption('brickset_username')} and \code{getOption('brickset_password')}.}
}
\value{
a data.frame with the subthemes.
}
\description{
Returns a table of sub-themes for a given theme with number of sets and years active.
}
