% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYears.R
\name{getYears}
\alias{getYears}
\title{Get a list of years for a given theme, with the total number of sets in each.}
\usage{
getYears(
  theme,
  key,
  username,
  password,
  userHash = login(username, password, key)
)
}
\arguments{
\item{theme}{the theme}

\item{username}{the Brickset username.}

\item{password}{the Brickset password.}

\item{userHash}{the hash used for a logged in user. See \link{login}.}
}
\value{
a data.frame with the years of a given theme.
}
\description{
Get a list of years for a given theme, with the total number of sets in each.
}
