% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUserHash.R
\name{checkUserHash}
\alias{checkUserHash}
\title{Check the Brickset API key.}
\usage{
checkUserHash(key, userHash, error = TRUE)
}
\arguments{
\item{key}{the API key}

\item{userHash}{the user hash returned from \link{login}.}

\item{error}{if TRUE, the function will throw an error. If FALSE, then the
function will return TRUE if the key check failed.}
}
\value{
TRUE if the API key is fine.
}
\description{
You can request an API key on the Brickset website here:
http://brickset.com/tools/webservices/requestkey
}
