% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUserHash.R
\name{checkUserHash}
\alias{checkUserHash}
\title{Check the Brickset API key.}
\usage{
checkUserHash(key = getOption("brickset_key"), userHash)
}
\arguments{
\item{key}{the API key}

\item{userHash}{the user hash returned from \link{login}.}
}
\value{
TRUE if the API key is fine.
}
\description{
You can request an API key on the Brickset website here:
https://brickset.com/tools/webservices/requestkey/
}
\details{
Brickset API documentation is available here:
https://brickset.com/article/52664/api-version-3-documentation/
}
