% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInstructions.R
\name{getInstructions}
\alias{getInstructions}
\title{Download list of instructions for given set.}
\usage{
getInstructions(setID, setNumber, key = getOption("brickset_key"), ...)
}
\arguments{
\item{setID}{the ID of the set (see \code{data(legosets)})}

\item{setNumber}{the set number from on the LEGO box}

\item{key}{the Brickset API key.}

\item{...}{other parameters passed to \code{\link{getUserHash}} including
the Brickset username and password if they are not available from
\code{getOption('brickset_username')} and \code{getOption('brickset_password')}.}
}
\value{
a data.frame with the instructions.
}
\description{
Brickset API documentation is available here:
https://brickset.com/article/52664/api-version-3-documentation/
}
