% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeyUsageStats.R
\name{getKeyUsageStats}
\alias{getKeyUsageStats}
\title{Get information about frequency of API usage for the given API key.}
\usage{
getKeyUsageStats(key = getOption("brickset_key"))
}
\arguments{
\item{key}{the API key}
}
\value{
a data.frame with the number of times the key has been used.
\describe{
\item{dateStamp}{The date}
\item{count}{The number of times the key was used for the given date}
}
}
\description{
Provides information about how frequently the API key has been used.
}
\examples{
\dontrun{
options(brickset_key = 'BRICKSET_KEY',
        brickset_username = 'BRICKSET_UERNAME',
        brickset_password = 'BRICKSET_PASSWORD')
getKeyUsageStats()
}
}
