% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReviews.R
\name{getReviews}
\alias{getReviews}
\title{Downloads Lego data.}
\usage{
getReviews(
  setID,
  key,
  username,
  password,
  userHash = login(username, password, key)
)
}
\arguments{
\item{key}{the Brickset key.}

\item{username}{the Brickset username.}

\item{password}{the Brickset password.}

\item{userHash}{the hash used for a logged in user. See \link{login}.}

\item{setId}{the ID of the set (see \code{data(legosets)})}
}
\value{
a data.frame with the reviews.
}
\description{
Downloads Lego data.
}
