% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSets.R
\name{getSets}
\alias{getSets}
\title{Downloads LEGO data from Brickset.}
\usage{
getSets(year, key = getOption("brickset_key"), ...)
}
\arguments{
\item{year}{the year of data to download.}

\item{key}{the Brickset key.}

\item{...}{other parameters passed to \code{\link{getUserHash}} including
the Brickset username and password if they are not available from
\code{getOption('brickset_username')} and \code{getOption('brickset_password')}.}
}
\value{
a data.frame with all sets from the given year.
}
\description{
Downloads LEGO data from Brickset.
}
