% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSets.R
\name{getSets}
\alias{getSets}
\title{Downloads Lego data.}
\usage{
getSets(
  year,
  key,
  username,
  password,
  userHash = login(username, password, key)
)
}
\arguments{
\item{year}{the year of data to download.}

\item{key}{the Brickset key.}

\item{username}{the Brickset username.}

\item{password}{the Brickset password.}

\item{userHash}{the hash used for a logged in user. See \link{login}.}
}
\value{
a data.frame with all sets from the given year.
}
\description{
Downloads Lego data.
}
