% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubthemes.R
\name{getSubthemes}
\alias{getSubthemes}
\title{Returns a table of sub-themes for a given theme with number of sets and years active.}
\usage{
getSubthemes(
  theme,
  key,
  username,
  password,
  userHash = login(username, password, key)
)
}
\arguments{
\item{theme}{the theme}

\item{username}{the Brickset username.}

\item{password}{the Brickset password.}

\item{userHash}{the hash used for a logged in user. See \link{login}.}
}
\value{
a data.frame with the subthemes.
}
\description{
Returns a table of sub-themes for a given theme with number of sets and years active.
}
