% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickset-package.R
\docType{data}
\name{legosets}
\alias{legosets}
\title{Lego sets from 1970 through 2020.}
\source{
\url{https://brickset.com}
}
\description{
Lego sets from 1970 through 2020.
}
\keyword{data,}
\keyword{lego}
