% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickset-package.R
\docType{data}
\name{legosets}
\alias{legosets}
\title{Lego sets from 1970 through 2024.}
\format{
A data.frame.
\describe{
\item{setID}{integer; 20420 unique values}
\item{number}{character; 18958 unique values}
\item{numberVariant}{integer; 25 unique values}
\item{name}{character; 17057 unique values}
\item{year}{integer; 55 unique values}
\item{theme}{character; 163 unique values}
\item{themeGroup}{character; 17 unique values}
\item{subtheme}{character; 1000 unique values}
\item{category}{character; 7 unique values}
\item{released}{logical; 2 unique values}
\item{pieces}{integer; 1543 unique values}
\item{minifigs}{integer; 34 unique values}
\item{bricksetURL}{character; 20420 unique values}
\item{rating}{numeric; 29 unique values}
\item{reviewCount}{integer; 64 unique values}
\item{packagingType}{character; 19 unique values}
\item{availability}{character; 11 unique values}
\item{agerange_min}{integer; 18 unique values}
\item{thumbnailURL}{character; 19365 unique values}
\item{imageURL}{character; 19365 unique values}
\item{US_retailPrice}{numeric; 173 unique values}
\item{US_dateFirstAvailable}{Date; 1090 unique values}
\item{US_dateLastAvailable}{Date; 2324 unique values}
\item{UK_retailPrice}{numeric; 224 unique values}
\item{UK_dateFirstAvailable}{Date; 1019 unique values}
\item{UK_dateLastAvailable}{Date; 2224 unique values}
\item{CA_retailPrice}{numeric; 188 unique values}
\item{CA_dateFirstAvailable}{Date; 851 unique values}
\item{CA_dateLastAvailable}{Date; 2005 unique values}
\item{DE_retailPrice}{numeric; 179 unique values}
\item{DE_dateFirstAvailable}{Date; 607 unique values}
\item{DE_dateLastAvailable}{Date; 1402 unique values}
\item{height}{numeric; 264 unique values}
\item{width}{numeric; 317 unique values}
\item{depth}{numeric; 307 unique values}
\item{weight}{numeric; 1192 unique values}
}
}
\source{
\url{https://brickset.com}
}
\description{
This dataset was built using the \code{\link{getSets}} function. For working
with the LEGO sets data frame this pre-built data is preferred as it will
minimize the API requests. Note that the only disadvantage is that the
\code{rating} and \code{reviewCount} may be out-of-date and inaccurate. The
remaining variables are relatively static.
}
\keyword{lego}
